gcp/use-case/second_part/requirements.txt
google-cloud-bigquery
google-cloud-pubsub
google-cloud-storage

